refactor: move write updater to utils
